% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_methods.R
\name{augment.CA}
\alias{augment.CA}
\title{Augment data with CA object}
\usage{
\method{augment}{CA}(x, data, for_columns = FALSE, ...)
}
\arguments{
\item{x}{An oebject of class \code{CA} from the FactoMineR package.}

\item{data}{The original data used to create the \code{CA} object.}

\item{for_columns}{augment columns instead}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{tibble} with columns containing the original data and additional columns with the row and column coordinates.
}
\description{
Augment data with information from a CA object
}
