% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca_methods.R
\name{augment.cc}
\alias{augment.cc}
\title{Augment data with CCA::cc object}
\usage{
\method{augment}{cc}(x, data, ...)
}
\arguments{
\item{x}{An object of class \code{CCA::cc} from the FactoMineR package.}

\item{data}{The original data used to create the \code{CCA::cc} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{tibble} with columns containing the original data and additional columns with the row and column coordinates.
}
\description{
Augment data with information from a CCA::cc object
}
