% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mca_methods.R
\name{augment.MCA}
\alias{augment.MCA}
\title{Augment data with MCA object}
\usage{
\method{augment}{MCA}(x, data, ...)
}
\arguments{
\item{x}{An object of class \code{MCA} from the FactoMineR package.}

\item{data}{The original data used to create the \code{MCA} object.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{tibble} with columns containing the original data and additional columns with the row and column coordinates.
}
\description{
Augment data with information from a MCA object
}
