% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca_methods.R
\name{autoplot.cca_processed}
\alias{autoplot.cca_processed}
\alias{autoplot.cca}
\title{Autoplot for cca_processed}
\usage{
\method{autoplot}{cca}(x, X, Y, type = c("all", "scree", "row", "col", "symmetric"), ...)
}
\arguments{
\item{x}{An object of class \code{cca} from the FactoMineR package.}

\item{type}{Type of plot to generate. Options are "all", "scree", "row", "col", "symmetric". Defaults to "all".}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot object or a combined grid of ggplot objects
}
\description{
Automatically generate plots for cca_processed object
}
