% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_methods.R
\name{colplot.CA_processed}
\alias{colplot.CA_processed}
\title{colplot for CA_processed}
\usage{
\method{colplot}{CA_processed}(augment_output, ...)
}
\arguments{
\item{augment_output}{Result of the augment function over CA}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot
}
\description{
colplot for CA_processed
}
