% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca_methods.R
\name{colplot.cca_processed}
\alias{colplot.cca_processed}
\title{colplot for cca_processed}
\usage{
\method{colplot}{cca_processed}(augment_output, ...)
}
\arguments{
\item{augment_output}{Result of the augment function over cca}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot
}
\description{
colplot for cca_processed
}
