% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mca_methods.R
\name{colplot.MCA_processed}
\alias{colplot.MCA_processed}
\title{colplot for MCA_processed}
\usage{
\method{colplot}{MCA_processed}(augment_output, ...)
}
\arguments{
\item{augment_output}{Result of the augment function over MCA}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot
}
\description{
colplot for MCA_processed
}
