% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_methods.R
\name{glance.CA}
\alias{glance.CA}
\title{Glance a CA object}
\usage{
\method{glance}{CA}(x, ...)
}
\arguments{
\item{x}{An object of class \code{CA} from the FactoMineR package.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{tibble} with statistics about computed parameters.
}
\description{
Global metrics about a computed CA object
}
