% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca_methods.R
\name{glance.cca}
\alias{glance.cca}
\title{Glance data with vegan::cca object}
\usage{
\method{glance}{cca}(x, ...)
}
\arguments{
\item{x}{An object of class \code{cca} from the FactoMineR package.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{tibble} with statistics about computed parameters.
}
\description{
Global metrics about a computed cca object
}
