% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_methods.R
\name{rowplot.CA}
\alias{rowplot.CA}
\title{rowplot for CA}
\usage{
rowplot.CA(augment_output, ...)
}
\arguments{
\item{augment_output}{Result of the augment function over CA}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot
}
\description{
rowplot for CA
}
