% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_methods.R
\name{rowplot.CA_processed}
\alias{rowplot.CA_processed}
\title{rowplot for CA_processed}
\usage{
rowplot.CA_processed(augment_output, ...)
}
\arguments{
\item{augment_output}{Result of the augment function over CA}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot
}
\description{
rowplot for CA_processed
}
