% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca_methods.R
\name{rowplot.cca_processed}
\alias{rowplot.cca_processed}
\title{rowplot for cca_processed}
\usage{
\method{rowplot}{cca_processed}(augment_output, ...)
}
\arguments{
\item{augment_output}{Result of the augment function over cca}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot
}
\description{
rowplot for cca_processed
}
