% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mca_methods.R
\name{rowplot.MCA_processed}
\alias{rowplot.MCA_processed}
\title{rowplot for MCA_processed}
\usage{
\method{rowplot}{MCA_processed}(augment_output, ...)
}
\arguments{
\item{augment_output}{Result of the augment function over MCA}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot
}
\description{
rowplot for MCA_processed
}
