% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rowplot}
\alias{rowplot}
\alias{colplot}
\title{Row plot}
\usage{
rowplot(augment_output, ...)

colplot(augment_output, ...)
}
\description{
Generic implementation of rowplot

Generic implementation of rowplot
}
