% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_methods.R
\name{rowplot_CA}
\alias{rowplot_CA}
\title{rowplot for CA}
\usage{
rowplot_CA(augment_output, ...)
}
\arguments{
\item{...}{Additional arguments (not used).}

\item{x}{Result of augment function over CA}
}
\value{
A ggplot
}
\description{
rowplot for CA
}
