% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_methods.R
\name{screeplot.CA_processed}
\alias{screeplot.CA_processed}
\title{screeplot for CA_processed}
\usage{
\method{screeplot}{CA_processed}(tidy_output, ...)
}
\arguments{
\item{tidy_output}{Result of tidy function over CA}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot
}
\description{
screeplot for CA_processed
}
