% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca_methods.R
\name{screeplot.cca_processed}
\alias{screeplot.cca_processed}
\title{screeplot for cca_processed}
\usage{
\method{screeplot}{cca_processed}(tidy_output, ...)
}
\arguments{
\item{tidy_output}{Result of tidy function over cca}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot
}
\description{
screeplot for cca_processed
}
