% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mca_methods.R
\name{screeplot.MCA_processed}
\alias{screeplot.MCA_processed}
\title{screeplot for MCA_processed}
\usage{
\method{screeplot}{MCA_processed}(tidy_output, ...)
}
\arguments{
\item{tidy_output}{Result of tidy function over MCA}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot
}
\description{
screeplot for MCA_processed
}
