% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_methods.R
\name{screeplot_CA}
\alias{screeplot_CA}
\title{screeplot for CA}
\usage{
screeplot_CA(tidy_output, ...)
}
\arguments{
\item{...}{Additional arguments (not used).}

\item{x}{Result of tidy function over CA}
}
\value{
A ggplot
}
\description{
screeplot for CA
}
