% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_methods.R
\name{symmetricplot.CA_processed}
\alias{symmetricplot.CA_processed}
\title{Symmetric plot for CA_processed}
\usage{
\method{symmetricplot}{CA_processed}(row_output, col_output, ...)
}
\arguments{
\item{row_output}{Result of the augment function over CA for rows}

\item{col_output}{Result of the augment function over CA for columns}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot
}
\description{
Symmetric plot for CA_processed
}
