% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca_methods.R
\name{symmetricplot.cca_processed}
\alias{symmetricplot.cca_processed}
\title{Symmetric plot for cca_processed}
\usage{
\method{symmetricplot}{cca_processed}(row_output, col_output, ...)
}
\arguments{
\item{row_output}{Result of the augment function over cca for rows}

\item{col_output}{Result of the augment function over cca for columns}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot
}
\description{
Symmetric plot for cca_processed
}
