% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mca_methods.R
\name{symmetricplot.MCA_processed}
\alias{symmetricplot.MCA_processed}
\title{Symmetric plot for MCA_processed}
\usage{
\method{symmetricplot}{MCA_processed}(row_output, col_output, ...)
}
\arguments{
\item{row_output}{Result of the augment function over MCA for rows}

\item{col_output}{Result of the augment function over MCA for columns}

\item{...}{Additional arguments (not used).}
}
\value{
A ggplot
}
\description{
Symmetric plot for MCA_processed
}
