% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mca_methods.R
\name{tidy.MCA}
\alias{tidy.MCA}
\title{Tidy a MCA object}
\usage{
\method{tidy}{MCA}(x, ...)
}
\arguments{
\item{x}{An object of class \code{MCA} from the FactoMineR package.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{tibble} with statistics about computed parameters.
}
\description{
statistics about a computed MCA object
}
