% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca_methods.R
\name{augment.cca}
\alias{augment.cca}
\title{Augment data with vegan::cca object}
\usage{
\method{augment}{cca}(x, data_X, data_Y, for_columns = FALSE, ...)
}
\arguments{
\item{x}{An object of class \code{vegan::cca} from the FactoMineR package.}

\item{data_X}{The original data used to create the \code{vegan::cca} object.}

\item{data_Y}{The original data used to create the \code{vegan::cca} object.}

\item{for_columns}{augment columns instead}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{tibble} with columns containing the original data and additional columns with the row and column coordinates.
}
\description{
Augment data with information from a vegan::cca object
}
